Add files via upload
Node is the node
PersonContact is the object class. Nothing is static anymore. Data is private.
PhoneBookMain is where the main method is. Commented out some test hard coded contact.
PhoneBookManager is all of the methods for adding, deleting, modifying, searching, and viewing.

You should be able to open it up on visual studio code through git hub. If you can't, just copy and paste.